Borg restoration content.
This removes any syndicate-related loadouts, adds 2 positions for both of the wasteland and 2 for oasis borgs, nerfed security loadout by replacing the disabler with a laser gun as the disabler is a pain in the ass.

Changed the descriptions of the borgs in oasis to be corrected as mayor and sheriff.

Removes any materials from engi because apparently it's infinite.

Mind you, the wasteland handys will either be killed by legion or just random wasties.